style: merge test_generate with test_data
